A hook to support the better-cond macro and all its features.
This is based on the hook first shown at

  https://github.com/clj-kondo/clj-kondo/blob/master/corpus/.clj-kondo/hooks/better_cond.clj

but it adds support for all the better-cond features and fixes (what
I think are) a couple bugs in the original (that caused it to fail
for me).

This version does take a shortcut in the expansion of :when-let and
:when-some that are good for linting but are not exactly equivalent
to the original macro.